refactor: icon파일 index.js로 모아서 re-export
