+++ frontend/src/assets/icons/index.js
@@ -1,0 +1,29 @@
+import KakaoIcon from './kakao.svg';
+import NaverIcon from './naver.svg';
+import ClockIcon from './clock.svg';
+import EyeIcon from './eye.svg';
+import EllipsisIcon from './ellipsis.svg';
+import HomeIcon from './home.svg';
+import ReviewIcon from './review.svg';
+import LoginIcon from './login.svg';
+import LogoutIcon from './logout.svg';
+import LogoIcon from './logo.svg';
+import LeftArrowIcon from './left-arrow.svg';
+import RightArrowIcon from './right-arrow.svg';
+import CloseIcon from './close.svg';
+
+export {
+  KakaoIcon,
+  NaverIcon,
+  ClockIcon,
+  EyeIcon,
+  EllipsisIcon,
+  HomeIcon,
+  ReviewIcon,
+  LoginIcon,
+  LogoutIcon,
+  LogoIcon,
+  LeftArrowIcon,
+  RightArrowIcon,
+  CloseIcon,
+};
