+++ frontend/src/components/Modal/Modal.js
@@ -1,7 +1,7 @@
 import PropTypes from 'prop-types';
+import { CloseIcon } from '../../assets/icons';
 import Frame from '../Frame/Frame';
 import { Container, modalFrameStyles, CloseButton } from './Modal.styles';
-import CloseIcon from '../../assets/icons/close.svg';
 
 const Modal = ({ children, showCloseButton, showShadow, onClickClose }) => (
   <Container>
