+++ frontend/src/components/SideBar/SideBar.js
@@ -1,11 +1,6 @@
 import { useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { useSnackbar } from 'notistack';
-import HomeIcon from '../../assets/icons/home.svg';
-import ReviewIcon from '../../assets/icons/review.svg';
-import LoginIcon from '../../assets/icons/login.svg';
-import LogoutIcon from '../../assets/icons/logout.svg';
-import LogoIcon from '../../assets/icons/logo.svg';
 import {
   Container,
   LogoContainer,
@@ -22,6 +17,7 @@ import {
   THEME_COLOR,
 } from '../../constants';
 import { getMyInfoAsync, logout as logoutAction } from '../../redux/authSlice';
+import { HomeIcon, LoginIcon, LogoIcon, LogoutIcon, ReviewIcon } from '../../assets/icons';
 
 const SideBar = () => {
   const dispatch = useDispatch();
